requirements.txt
asgiref==3.4.1
Django==3.2.8
djangorestframework==3.12.4
pytz==2021.3
sqlparse==0.4.2
djangorestframework-simplejwt==5.0.0
pyjwt==2.3.0
python-decouple==3.5
psycopg2==2.9.1
black==21.10b0
django-rest-framework-datatables-filter==0.0.2
flake8==4.0.1
psycopg2-binary==2.9.2
dj-database-url==0.5.0
gunicorn==20.1.0
whitenoise==5.3.0
coverage==6.1.2